fix: check if player has a heavy item on item drop

+++ Api/src/Status/Listener/EquipmentSubscriber.php
@@ -2,6 +2,7 @@
 
 namespace Mush\Status\Listener;
 
+use Mush\Equipment\Entity\GameItem;
 use Mush\Equipment\Event\EquipmentEvent;
 use Mush\Equipment\Event\TransformEquipmentEvent;
 use Mush\Player\Entity\Player;
@@ -91,7 +92,11 @@ class EquipmentSubscriber implements EventSubscriberInterface
         $player = $equipment->getHolder();
         if ($player instanceof Player &&
             $player->hasStatus(PlayerStatusEnum::BURDENED) &&
-            $equipment->hasStatus(EquipmentStatusEnum::HEAVY)
+            $equipment->hasStatus(EquipmentStatusEnum::HEAVY) &&
+            $player->getEquipments()->filter(function (GameItem $item) {
+                return $item->hasStatus(EquipmentStatusEnum::HEAVY);
+            })->count() >= 1
+
         ) {
             $this->statusService->removeStatus(PlayerStatusEnum::BURDENED, $player, $reasons, $time);
         }
